requirements.txt
discord.py
discord.py[voice]
python-dotenv
wavelink
PyNaCl
motor
pymongo